telegram_resume_bot/requirements.txt
python-telegram-bot
python-docx
docx2pdf
pytz